Do not change current index of mapped model if the mapped one is invalid
Assume viewA and viewB, showing modelA  and modelB. modelB is a subset of modelA,
viewA and viewB shares a selection model through KLinkItemSelectionModel. When
current index of viewA does not exist in viewB, do not set viewB current index
to invalid, as it would in turn set viewA index to invalid.

REVIEW: 112756