KPluginInfo rework for KPluginTrader
This patch prepares KPluginInfo for usage with KPluginTrader (to be
submitted in a separate patch). It basically makes KPluginInfo's API a
little bit more like KService by adding a property(QString) accessor to
the plugin info.

Part of this patch, and much of its churn, is the internal change from
independent QStrings and QStringLists to a QVariantMap. This is how the
metadata comes in from KPlugin*, and it allows us to make properties
accessible by name.

There's also a fair bit of moving from QLatin1String to QStringLiteral
in there, most of these lines needed changes anyway. Additionally, the
keys of properties are now shared in the d-pointer.

This change is source compatible to the old version.

REVIEW:112660