Make KTraderParse* fit to search KPluginInfo::List
This patch allows search through KPluginInfo::List, additionally to
KService::List. Together with the other patches from my set It makes it
possible to search through the plugin's (JSON-based) metadata, using our
constraint syntax.

It adds a ctor that takes a KPluginInfo::List& and adds the necessary
branches to look within either the KService::List or the
KPluginInfo::List.

REVIEW:112679