Start adopting QCollator
All the porting was quite straightforward I'd say, the only weird part is
that I removed some tests of the KStringHandler tests. There are 2 kind of
tests disabled:
- The ones that say that "A" == "a" in case of Qt::CaseInsensitive. ICU is
deterministic and it will decide itself which one goes in, so the test
doesn't make sense anymore.
- There's a mention to the 237788 bug where in some cases our former
algorithm wouldn't be deterministic. This doesn't apply anymore, but also
now ICU takes care about it now, so there's little point of keeping unit
testing it.
I decided to leave the unit test because it might be useful eventually
(although note that it was not being compiled so far). In any case we
probably want it out.

In any case, the rest seems straightforward enough. I didn't concentrate
on performance though, in some cases we'll want to use the
QCollatorSortKey.

REVIEW: 112717