Fix KIO::listRecursive going up the folder hierarchy (!)
This was a side effect of QUrl::setPath() now normalizing the path:
setPath("/my/dir/..") used to give fileName() == "..", but now it
gets normalized to "/my/" so fileName() is empty.
The code was checking fileName for "." and ".." explicitely, hence the bug.

Fixed by skipping the roundtrip via QUrl (which was a bit stupid anyway),
but I wonder how many of these we'll see. Maybe the earlier version of the
patch (cleaning duplicate slashes, but not "." and "..") was safer.
CCMAIL: thiago@kde.org