Prepare KDNSSD for moving
Prepares the module to be moved to its own tier.
There's a rather uncommon case with this module because there's a
dependency against KConfig, but it only applies to when it's compiled with
the mdnsd backend (only when there's no Avahi but mDns instead). I guess
this makes it tier 2, but I'm not all that sure about that.

REVIEW: 112900