Try to compile a basic Qt application before the -fvisiblity=hidden check
If there is a problem with finding the Qt includes, then
FindKDE4Internal tells the user that Qt is not compiled correctly, and
it is not straight forward to realise that it is a false alarm. This
most likely happens for cmake < 2.8.11.

Introduce a simple check for compiling a barebone Qt application before
attempting to check for -fvisibility=hidden. This way, the error message
states that Qt has not been found correctly, preventing confusion.

REVIEW: 112969