kcookiejar: Fix parsing of weird dates
This fixes kcookiejartest, which was failing to parse the date from the
Y2K38 check.

The test started failing on build.kde.org when the Qt version there was
switched to the "stable" branch, which seems to parse things differently
than the version of Qt build.kde.org used to run.

The code was ported from KDateTime::fromString() to
QDateTime::fromString(), whose syntax for date formatting is different,
but the format was not updated.

REVIEW: 113076