Comment usage of HAVE_X11 in kded
xlib is used to talk with kplash and inform of the loading state. In KF5
plasma2 we will have a new startup system that for sure won't be using
X11, or making synchronous things so it is certant that we don't need
this code.

I'm leaving the code there to make sure that it is removed only when
the new startup system is in place.