Adapt to the qt5/stable refactoring on the event dispatcher
The ownership of the event dispatcher has been changed in platform
plugins. We're not supposed to create socket notifiers anymore in ctors
for instance. Since we did indirectly through D-Bus, we have to delay
that now so that it's created only when the event loop is running.

It fixes the issues I was having with kded not talking on the bus at all
when the platform theme was installed.