Reduce warnings noise when including wtf/Platform.h in KJS
Modifies how features are checked. This file has some pre-processor macros
that are used to checking for OS features. These are defined with "1" if
acceptable and undefined if the feature is not present. These macros are
being checked with #if, so the pre-processor will complain whenever it's
asked about one of the variables that haven't been defined.

This patch adds a defined() first, before checking the value, so the
pre-processor doesn't need to consider the variable value, if it's not
present.

This change makes it possible to read the compilation logs from modules
that use KJS.

REVIEW: 113173