Preparing KTimeZone for behavior changes in qt5/stable
Two changes because QDateTime is becoming more picky as it gets more
timezone aware itself in qt5/stable. Now QDateTime instances for which
we should return InvalidOffset are outright invalid, so let's return
that instead of 0 in the error case (makes more sense to me too). And
for the timezone during the tests, they were written assuming they'd be
run in the london timezone, and now it matters that we make sure of
that, otherwise QDateTime won't behave as expected.

I think we don't see the failures yet on build.kde.org because it's
slightly lagging behind or because the system is in the right timezone.