kdeui: Streamline KPasswordDialog.
A user reported a couple of years ago that the "Anonymous login"
checkbox is badly placed in KPasswordDialog, appearing in between the
user and password fields and sticking out like a sore thumb.

Instead of just moving the checkbox, I took the opportunity to make the
U/I match the underlying code. Since "Anonymous" mode makes the existing
U/I set both username and password fields to be read-only, it's really a
separate authentication scheme and should be treated as such, especially
since it is hard to tell visually that the Anonymous checkbox is
responsible for disabling the line edits above and below it.

The existing code relies somewhat on the widgets that are present in the
U/I file, so instead of trying to merge the new radio buttons with the
existing widgets I separated out the "anonymous" radio button and its
new counterpart, and left the rest of the existing widgets parented
under a hidden widget to minimize code churn and the risk of introducing
bugs.

Tested with a handy FTP server, and with the included KPasswordDialog
test suite.

BUG:272313
FIXED-IN:4.12
DIGEST:Improvements to KPasswordDialog
GUI: