Add include/KDE to users of KIOCore.
Note that this really only hides the problem. There are several problems
with the KF5 includes, and several patches in review which also don't
fully solve the problem.

What is needed is to figure out

* Where are headers installed to?
* How are users supposed to use them?
* What include paths do users need to use?

For example, if kfoo installs

 include/kfoo/job.h
 include/kfoo/kfoo_export.h

and users are supposed to use:

 .#include <kfoo/job.h>

but job.h #includes <kfoo_export.h>

then users need both

 include/
 include/kfoo/

when compiling. That seems to be the case at some points in KDE
frameworks, and it is more-likely accidental than intentional.

That should be reviewed.

CCMAIL: kde-frameworks-devel@kde.org