Use a better hack for dealing with gperf's unhelpful inlining
gperf (in ANSI-C mode) attempts to inline functions in gnu90 mode, which
*only* generates inline versions.  This causes linker errors if the
compiler decides not to use inlining.

Previously, there was a hack (claiming to prevent warnings) that
disabled __inline keyword, which fixed the linking issue but caused
another warning (about ignored attributes) to be emitted because the
__gnu_inline__ attribute was still there, and that doesn't work without
one of the normal inline keywords.

This version leaves __inline alone, but prevents the use of
__gnu_inline__.

REVIEW: 113430