Ensure neither queue nor thread keep a reference to a previous job.
See comments on JobTests::JobsAreDestroyedAfterFinish:

Verify that neither the queue nor the thread keep a reference to the job
after completing it.

This is necessary because user-allocated objects like queue policies may
be registered with the jobs. If the jobs stick around until the thread
or queue are deleted, the user-allocated objects may have gone out of
scope or been deleted already, causing potential errors. From
ThreadWeaver's point of view, a job seizes to exist once the processing
thread asks for the next job.