First version of a convenience API for queueing and creating jobs.
The API of the queue is supposed to be verbose and minimal. The purpose
of the convenience API is to complement the queue API with simple and
terse ways to create and queue jobs. It provides provides templates for
creating JobPointers from Jobs and functors, for queueing JobPointers,
Jobs and Functors, and a stream like API for queueing multiple jobs in
one statement.

With these extensions enqueueRaw and dequeueRaw are obsolete and have
been removed.