New KColorSchemeManager to support changing color scheme in app
This class is inspired by functionality offered by e.g. Krita and
Digikam to allow the user to select a different color scheme for the
application.

This manager simplifies this task and also ensures that the required
property on QApplication is set, so that a QStyle can pass the scheme
to the window manager/compositor for the windows of the application.

REVIEW: 113685