Unbreak build
Moc was failing on the KDE4SUPPORT_DEPRECATED macro. I initially tried to
just wrap it in a "#ifndef KDE_NO_DEPRECATED" block, but it still did
not like it, so I removed the KDE4SUPPORT_DEPRECATED.

What is the proper way to mark a signal deprecated?

CCMAIL: kde-frameworks-devel@kde.org