Unit tests for NETRootInfo and NETWinInfo
The unit test starts an Xvfb in the initTestCase and closes it again
when the tests end. The display number is chosen by random in [1,98].
0 is skipped as that is the normal X server and 99 is skipped as that's
what KDE's CI infrastructure already uses.

An explicit Xvfb is started instead of using an already running
X Server to not interfere with already running window managers or
left-over states from other tests.

It would be even better to have one Xvfb instance per test method,
but this is not possible as the atoms are only resolved once and
thus NETRootInfo is indirectly bound to a Display although the API
kind of implies that it would be able to use different NETRootInfo
for different Displays.

Coverage of the unit tests is not yet complete.