Make it possible to avoid excessive warnings.
Inline implmenentations, implicit conversion in method parameteres etc
cause many deprecation messages for users. The backtrace of those messages
does not lead back to user code, but is bound to the inline
implementation. The message is emitted because of an include. Make it possible
to avoid such value-free messages, and leave a #warning instead.