Properly erase KSartupInfoIds in remove_startup_info_internal
If there were multiple KStartupInfoIds which were removed at the same
time, it was possible that the map got changed while being iterated,
thus the data corrupted and we hit a crash later on. This changes to
using erase() on the map to get a valid iterator again.

In order to get the added unit test run in a sensible time frame, an
environment variable KSTARTUPINFO_TIMEOUT is added to control the
startupinfo's timeout in a better way.

REVIEW: 113877