Save memory by sharing QStrings in UDSEntry
If many UDSEntries are received from a kioslave, some of the QString
fields will often be the same for many entries in a row (for example,
the user name and the group). In this case, we can save memory by making
use of QString's implicit sharing feature for these strings.

Moreover, this commit adds a unit test that verifies that storing some
UDSEntries to a QDataStream and re-loading them works as expected.

REVIEW: 113591