Workaround for dependency issues when generating .json from .desktop files
There is currently no way to declare a target as a dependency of the automoc
target, so it is not possible to ensure .json files are generated from .desktop
files before moc is run. To work around this, generate .json files at configure
time.

This is a workaround because it means changes to .desktop files won't be applied
to .json files when building, one will need to re-run cmake.

This change adds a test to ensure the CMake macro works well. It is declared as
an external project because it needs desktoptojson to be installed to work
correctly.

REVIEW: 113949