KDirWatch: fix stop+restart scan for all backends
* The docs says no signal will be emitted, so disable the code that
was remembering the events in a list of "pending" events.

* Fix behavior difference between inotify and other backends, which would
even emit "created" when restarting the watching (completely wrong!)

* Adjust code that would trust QFSW rather than compare mtimes: clearly
it was written as a workaround for Windows ("adding files to directories
doesn't change the mtime on Windows"), so on Unix keep the mtime comparison
to avoid spurious signals when restarting the watching. The OS-dependent
behavior isn't great though.

So even though this makes KDirWatch behave more as documented
(no signal while watching is stopped), it still emits signals when
watching is restarted, on Windows. Could be fixed for individual files,
though, at least.

CCMAIL: amantia@kde.org