Add manual test app for enableBlurBehind.
Added a simple app to test enableBlurBehind. It shows a semi-transparent
window, with buttons to enable or disable the blur. It can enable blur
for the whole window, or for a region below the buttons (rectangle or
circle).

Known issue: If blurring a region, currently it won't update the blur area
when the window is resized.

REVIEW:114050