Make KLimitedIODevice::bytesAvailable return the numbers of bytesAvaliable to read
At the moment KLimitedIODevice::bytesAvailable is returning at least the size of the file, so QIODevice::atEnd never returns true because it thinks there are more bytes to read. This makes that if you feed an "invalid" svg like "<?xml version="1.0"?>" to the QImageReader it infinite loops believing there will be more stuff to read but then read() always returns 0 but then bytesAvailable says there are more and it stays there forever.

BUGS: 328182
FIXED-IN: 4.11.5
REVIEW: 114226