Add include for klocalizedstring in klocale.h
A common porting error is the lack of the KLocalizedString include.
In KDE4 times this was also included through klocale.h. Many files
have an include to klocale for historic reasons and do not include
klocalizedstring at all.

By adding this include to klocale.h we can significantly decrease
the initial "get-it-compile-pain".

REVIEW: 114200