Properly set the event mask on the windows watched by KWindowSystem
KWindowSystem tried to only change the event mask for windows not
belonging to the same process by using QWidget::find to test whether
the window belongs to the same process or not. This does no longer
work in a QWindow world. If the window belongs to a Q(Quick)Window
QWidget::find returns a nullptr just like for windows of another
process. This caused the XSelectWindow to overwrite the event mask
used by Qt and thus for example mouse events were no longer recieved
at all.

With this change the event mask is updated for all windows, no matter
whether it's an own or a foreign window. But the event mask currently
installed on the window is preserved, so Qt still gets the events it
expects.

REVIEW: 114396