Do not assume every items have the same height
1. the offset addjust in `KCategorizedView::indexAt` was a no-op (it
operates on a temporary variable and is not needed).
2. KCategorizedView::indexAt (effectively) assumes all items has the
same height when doing bsearch and therefore failed to handle some cases
when the text takes multiple lines as shown in the bug report.

This patch removes the no-op and add special check for items in the same
row on the left (or on the right for RightToLeft layout) in order to
determine which way the bsearch should go.

REVIEW: 113969