KJS: treat time correctly as localtime in the Date() constructor
Commit 48dd1fc50277b861b49613c5f46b6f8b10cc932d introduced a bug:
It factors out the code to convert a time specification to an actual
time value into the function makeTimeFromList().
But in that function makeTime(&t, ms, true) is called, whereas the
original code called makeTime(&t, ms, false) for the constructor.

This patch fixes it by passing the "utc" parameter to makeTime() instead
of "true".

BUG: 335556
FIXED-IN: 3.13.3
REVIEW: 118947