Fix crashes in KUrlNavigator that are caused by nested event loops
KUrlNavigator opens menus with exec() in a few places, and accesses
member variables or pointers to children after that. This can cause
crashes if the object has been deleted inside the nested event loops.

This is fixed by using QPointers to detect if an object was deleted
already, and return immediately in that case.

I'll forward-port to KF5 in a few days.

BUG: 293863
REVIEW: 118858
FIXED-IN: 4.13.3