Make FindPyQt4 work with PyQt's new build system.
Since PyQt 4.10, PyQt.pyqtconfig is deprecated and not available unless
PyQt is built using the old configure script.

There is no direct replacement for it, as PyQt's new build system does
not provide as much information as before by design. Luckily, most of
the variables we are interested in can be obtained from PyQt's QtCore
module itself even if its old build system is used.

The only exception is pyqt_sip_dir, which cannot be determined at all if
pyqtconfig is not available. In this case, the most we can do is guess
the default path like QScintilla2 does, and fail if it does not exist.
The user then needs to specify it manually via CMake with something like
-DPYQT4_SIP_DIR=/usr/share/sip/PyQt4. To this effect, all variables set
by FindPyQt4.cmake have been made cache variables, which means their
values can be overriden by the user, thus ignoring the contents read via
FindPyQt.py.

BUG:      337462
FIXED-IN: 4.14.0
REVIEW:   119302