Sometimes when the system has high IO load (e. g. copying huge files or compiling and linking) some KDE apps become unresponsive. This seems to be caused by KDE's internal sockets created in /tmp which is normally on the same block device as the ongoing IO traffic.
With this patch the sockets are created in the folder XDG_RUNTIME_DIR points to. On "current" 
systems this folder is mounted with the tmpfs file system. So the sockets aren't affected by 
high IO load anymore.

In frameworks QStandardPaths is used which does the same, so there is no need to port it.

REVIEW:120492