Make quoteArgs quote urls with QChar::isSpace()
As David says in the review this is more a workaround than the real fix,
the real fix would involve modifying splitArgs so that it does not split where it is not needed.

But this is an easier, less impact fix since i don't really know if all shells only split on *real* space or not.

Acked by mpyne

Backported from kcoreaddon b746dcf93239e93eafeedfaeff53e1ac387f5ca1

BuGS: 343380
FIXED-IN: 4.14.6