Remove policy settings from FindKDE4Internal.
At this point, the ones which are set here are all set to NEW, except
CMP0011.  The point of CMP0011 here is to make the policy settings
be used by consumers.  All consumers need to gain a use of
the cmake_minimum_required command now anyway to satisfy CMP0000, so
just remove the call in the internal file.