Fix KDirListerCache crash with two listers for an empty dir in the cache.
The wrong optimization if (!itemU->lstItems.isEmpty()) meant that if
the cached dir is empty, we didn't create a CachedItemsJob. If there is
also no ListJob happening, then the lister would just be forever in the
listersCurrentlyListing list, which would lead to crashes later on.

The user-visible bug was hard to trigger (needed autofs enabled, I suppose to
trigger [1], although not sure why it's related), but after enabling DEBUG_CACHE,
simply clicking on the trash place entry in dolphin would trigger an assert.

[1] "KDirListerCache::forgetDirs: Not adding a watch on "file:///" because it contains a manually mounted subdir"

BUG: 278431
FIXED-IN: 4.14.17