Support SVG too
Konqueror's about page doesn't show icons if the breeze icon theme is
used.
The reason is that khtml doesn't support SVG images.
This patch is supposed to fix it by allowing to use Qt's support in that
case.

This is a backport of commit ed5c039f2d1996545b95a24e9bff4264144f9f6d
in khtml

FIXED-IN: 4.14.35
CCBUG: 355872