Security: remove support for $(...) in config keys with [$e] marker.
It is very unclear at this point what a valid use case for this feature
would possibly be. The old documentation only mentions $(hostname) as
an example, which can be done with $HOSTNAME instead.

Note that $(...) is still supported in Exec lines of desktop files,
this does not require [$e] anyway (and actually works better without it,
otherwise the $ signs need to be doubled to obey kconfig $e escaping rules...).

Thanks to Fabian Vogt for testing.

(This is a backport of KDE Frameworks 5 kconfig patch to kdelibs)

Differential Revision: https://phabricator.kde.org/D22989